Task 6

+++ 0x00-hello_world/6-size.c
@@ -6,12 +6,12 @@
  */
 int main(void)
 {
-	printf("Size of a char: %lu byts(s)\n", sizeof(char));
-	printf("Size of an int: %lu byts(s)\n", sizeof(int));
-	printf("Size of a double: %lu byts(s)\n", sizeof(double));
-	printf("Size of a long int: %lu byts(s)\n", sizeof(long int));
-	printf("Size of a long long int: %lu byts(s)\n", sizeof(long long int));
-	printf("Size of a float: %lu byts(s)\n", sizeof(float));
+	printf("Size of a char: %lu byts(s)\n", (unsigned)sizeof(char));
+	printf("Size of an int: %lu byts(s)\n", (unsigned)sizeof(int));
+	printf("Size of a double: %lu byts(s)\n", (unsigned)sizeof(double));
+	printf("Size of a long int: %lu byts(s)\n", (unsigned)sizeof(long int));
+	printf("Size of a long long int: %lu byts(s)\n", (unsigned)sizeof(long long int));
+	printf("Size of a float: %lu byts(s)\n", (unsigned)sizeof(float));
 
 	return (0);
 }
